fix: test refactor
